+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/core/EncodeTransferMedataScgEnhancedPlugin.java
@@ -82,7 +82,7 @@ public class EncodeTransferMedataScgEnhancedPlugin implements EnhancedPlugin {
 
 		contextToHeaderInterceptorList.forEach(ContextToHeaderInterceptor::beforeContextToHeader);
 
-		Map<String, String> customMetadata = metadataContext.getCustomMetadata();
+		Map<String, String> customMetadata = metadataContext.getTransitiveMetadata();
 		Map<String, String> disposableMetadata = metadataContext.getDisposableMetadata();
 		Map<String, String> applicationMetadata = metadataContext.getApplicationMetadata();